+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.1.4 - Unreleased
 
 - Added 32-bit bindings in -sys crate.
+- Added Stepper motor support. (missing async functionality)
 
 ## [v0.1.3](https://github.com/fpagliughi/phidget-rs/compare/v0.1.2..v0.1.3)  - 2024-03-10